+++ binding.gyp
@@ -7,21 +7,25 @@
             "cflags!": [ "-fno-exceptions" ],
             "cflags_cc!": [ "-fno-exceptions" ],
             "cflags_cc" : ["-std=c++20"],
-            "sources": [ "addon.cc" ],
+            "sources": [],
             "include_dirs": [
+                "src",
                 "<!@(node -p \"require('node-addon-api').include\")"
             ],
             "dependencies": [],
             'defines': [ 'NAPI_DISABLE_CPP_EXCEPTIONS' ],
             'conditions': [ # Initials checks on tray/Makefile
                 ["OS=='linux'", {
+                    "sources": ["src/linux/tray.cc"],
                     "cflags+": ["-DTRAY_APPINDICATOR=1", "<!@(pkg-config --cflags appindicator3-0.1)"],
                     "ldflags+": ["<!@(pkg-config --libs appindicator3-0.1)"],
                 }],
                 ["OS=='win'", {
+                    "sources": ["src/win/tray.cc"],
                     "defines":["TRAY_WINAPI=1"]
                 }],
                 ["OS=='mac'", {
+                    "sources": ["src/mac/tray.cc"],
                     "cflags+": ["-DTRAY_APPKIT=1"],
                     "dflags+": ["-framework Cocoa"]
                 }]
